Remove the ability to add helpers.
Initial plan was to have this configuration such that we can add `Rails.application.routes.url_helpers` to ActiveWaiter.
Adding that was only helpful in clearing errors when a parent app layout was used with routes originating from parent app.
However, the URL would still look like `host/active_waiter/.../whatever`, hence it didn't really solve all our problems.
To have the routes appear properly, you will still have to rename the routes used in the parent layout to `main_app.url`.

In light of that, this configuration is not very useful, so removing it for now.